pyproject.toml
[project]
name = "bikeheadvr"
version = "0.1.0"
description = "SteamVR overlay prototype for controllerless VRChat bike locomotion"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openvr>=2.12.1401",
    "pillow>=12.1.1",
    "platformdirs>=4.5.0",
    "pyside6>=6.10.0",
    "pyglet>=2.1.13",
    "python-osc>=1.9.3",
]

[project.scripts]
bikeheadvr = "bikeheadvr.desktop:main"
bikeheadvr-cli = "bikeheadvr.app:cli_main"

[build-system]
requires = ["uv_build >= 0.10.10, <0.11.0"]
build-backend = "uv_build"

[tool.uv.workspace]
members = [
    "bikeheadvr",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.15.7",
    "ty>=0.0.24",
]
build = [
    "pyinstaller>=6.16.0",
]

[tool.ruff]
line-length = 88
target-version = "py313"
src = ["src"]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.pytest.ini_options]
addopts = "-p no:cacheprovider"
testpaths = ["tests"]
